02-twitter | refactoring, NweetFactory.js 생성 #67
